requirements.txt.txt
streamlit
edge-tts
google-generativeai
pandas
openpyxl